chore: 재배포 (#482)
* refactor: update cron task names and add key-based filtering in getCronLogs

- getCronLogs 메소드에 key 파라미터를 추가하여 taskName을 기준으로 필터링 가능하도록 수정
- Cron 작업의 로그 태스크 이름을 간결하게 변경

* refactor: update cron task names and add key-based filtering in getCronLogs

- getCronLogs 메소드에 key 파라미터를 추가하여 taskName을 기준으로 필터링 가능하도록 수정
- Cron 작업의 로그 태스크 이름을 간결하게 변경

* feat: store database implementation

* feat: theme crud

* feat: item crud

* feat: view and purchase themes

* feat: view and purchase items

* feat: activate themes and items

* feat: Add layout information to user basic information and modify caching

* doc: swagger update

* doc: swagger update

* doc: swagger update

* fix: problem with published posts not being viewed properly due to paging issues

* test: Added console log to resolve issues that occur when using rt

* test: Added console log to resolve issues that occur when using rt

* feat: middleware is added to add request request identifier for testing.

* feat: middleware is added to add request request identifier for testing.

* fix: replace app.enableCors with express cors middleware for custom headers

* fix: replace app.enableCors with express cors middleware for custom headers

* fix: add X-Access-Token to Access-Control-Expose-Headers

* feat: Add basic theme creation logic

* refactor: guard refactoring and improving duplicate handling logic

* fix: remove tag limit

* refactor: item and theme purchase logic by applying Redis distributed locks and transactions

* refactor: Add retry logic to method

* refactor: Added name duplication check logic when saving and updating stories

* fix: 키워드 타입 검사 추가

* fix: 키워드 타입 검사 추가

* chore: Update dependency package versions to latest

* chore: 의존성 패키지 업데이트

* chore: 의존성 패키지 업데이트

* chore: 의존성 패키지 업데이트

* chore: eslint updated

* fix: redis connection settings in redlock

* fix: 에세이 조회 분산락에 지수 백오프 적용

* doc: edit readme

* chore: edit eslint config

* fix: 땅에 묻기 기능 추가로 인한 기존 로직 변경

* fix: 땅에묻기 필수 좌표 분기 추가

* refactor: geometry 타입 변경 밑작업

* fix: 에세이 작성시 위도,경도 데이터 geometry로 변환

* feat: 사용자 위치 기반 주변 에세이 알림 추가

* feat: 사용자 위치 기반 주변 에세이 알림 추가

* fix: 장치 설정 검색을 일괄 처리하여 성능 최적화

* fix: 관리자 검색 API에 대한 페이지네이션 추가

* feat: 유저 및 에세이 유사도 검색 추가

* feat: 사용자 위치 기반 주변 에세이 리스트 조회

* refactor: 에세이 상세조회 api에 위치기반 에세이 조회 페이지 기능 추가

* test: alert service 테스트 코드 개선

* test: auth service 테스트 코드 개선

* test: badge service 테스트 코드 개선

* test: bookmark service 테스트 코드 개선

* test: burial service 테스트 코드 개선

* test: essay service 테스트 코드 개선

* fix: 에세이 상태 타입 리네임

* draft: essay 테스트코드 개선

* fix: 스토리 생성 후 응답본문 추가

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 마이그레이션 수정

* fix: 스키마 강제동기화 및 함수셋업

* fix: 스키마 강제동기화 및 함수셋업

* fix: 데이터베이스 초기화

* fix: 데이터베이스 인덱스 및 함수 세팅

* fix: 데이터베이스 인덱스 및 함수 세팅

* fix: PageType Enum 유효성 검사 업데이트 및 토큰 만료 확인 추가

* fix: 스토리생성, 수정에 사용하는 에세이리스트 조회 api의 쿼리 수정

* fix: 스토리생성, 수정에 사용하는 에세이리스트 응답구조에 현재 스토리 이름 추가

* fix: 관리자 리스트 조회시 필수쿼리 옵셔널 파이프 설정

* feat: 새로운 스토리에세이 리스트 조회 경로 추가

* refactor: 스토리 에세이리스트 조회 로직 변경

* fix: 에세에 소유주 분기처리 및 AOS 점진적 업데이트를 위한 검증로직 주석처리

* fix: 에세이리스트 스토리 조회시 응답본문에 스토리 이름 추가

* draft: 에세이 도메인 테스트코드 개선

* draft: 에세이 도메인 테스트코드 개선

+++ src/modules/essay/essay.service.ts
@@ -1,11 +1,4 @@
-import {
-  forwardRef,
-  HttpException,
-  HttpStatus,
-  Inject,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { forwardRef, HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
 import { Transactional } from 'typeorm-transactional';
 import Redis from 'ioredis';
 import { InjectRedis } from '@nestjs-modules/ioredis';
@@ -196,7 +189,7 @@ export class EssayService {
     pageType: PageType,
     page: number,
     limit: number,
-    storyId: number,
+    storyId?: number,
   ) {
     let { essays, total } =
       // todo 안드로이드 다음 릴리즈 적용시 해제
@@ -207,12 +200,16 @@ export class EssayService {
 
     const totalPage: number = Math.ceil(total / limit);
 
+    let currentStoryName: string;
+    if (storyId !== undefined)
+      currentStoryName = await this.utilsService.findStoryNameInEssays(essays);
+
     essays.forEach((essay) => {
       essay.content = this.utilsService.extractPartContent(essay.content);
     });
     const essayDtos = this.utilsService.transformToDto(SummaryEssayResDto, essays);
 
-    return { essays: essayDtos, total, totalPage, page };
+    return { essays: essayDtos, total, totalPage, page, currentStoryName };
   }
 
   async getTargetUserEssays(userId: number, storyId: number, page: number, limit: number) {
@@ -356,20 +353,22 @@ export class EssayService {
 
     if (lock) {
       try {
-        const newTrendScore = await this.calculateTrendScore(essay);
+        const newTrendScore = await this.utilsService.calculateTrendScore(essay);
         const aggregate = await this.updateAggregateData(essay, newTrendScore);
 
         await this.redis.set(`aggregate:${essay.id}`, JSON.stringify(aggregate), 'EX', 300);
       } finally {
-        await this.redis.del(lockKey);
+        if (lock) {
+          await this.redis.del(lockKey);
+        }
       }
     } else {
-      console.log(`락 획득 실패: ${essay.id}`);
+      throw new HttpException(`락 획득 실패: ${essay.id}`, HttpStatus.TOO_MANY_REQUESTS);
     }
   }
 
-  async acquireLock(lockKey: string): Promise<string | null> {
-    const lockTimeout = 30;
+  async acquireLock(lockKey: string) {
+    const lockTimeout = 100;
     const maxAttempts = 5;
     const baseRetryDelay = 10;
     let lock: string | null = null;
@@ -383,26 +382,7 @@ export class EssayService {
         await new Promise((resolve) => setTimeout(resolve, retryDelay));
       }
     }
-    return lock;
-  }
-
-  async calculateTrendScore(essay: Essay) {
-    const incrementAmount = 1;
-    const decayFactor = 0.995;
-    const currentDate = new Date();
-    const createdDate = essay.createdDate;
-    const daysSinceCreation =
-      (currentDate.getTime() - new Date(createdDate).getTime()) / (1000 * 3600 * 24);
-
-    let newTrendScore = essay.trendScore;
-    if (daysSinceCreation <= 7) {
-      newTrendScore += incrementAmount;
-    } else {
-      const daysSinceDecay = daysSinceCreation - 7;
-      newTrendScore = essay.trendScore * Math.pow(decayFactor, daysSinceDecay);
-      newTrendScore = Math.floor(newTrendScore) + incrementAmount;
-    }
-    return newTrendScore;
+    return !!lock;
   }
 
   async updateAggregateData(essay: Essay, newTrendScore: number): Promise<Aggregate> {
@@ -475,6 +455,8 @@ export class EssayService {
         storyId,
         essay.createdDate,
       );
+    } else {
+      throw new HttpException('잘못된 요청입니다.', HttpStatus.BAD_REQUEST);
     }
 
     previousEssay.forEach((essay) => {
@@ -488,6 +470,9 @@ export class EssayService {
   @Transactional()
   async deleteEssay(userId: number, essayId: number) {
     const essay = await this.essayRepository.findEssayById(essayId);
+
+    if (!essay) throw new HttpException('에세이를 찾을 수 없습니다.', HttpStatus.NOT_FOUND);
+
     await this.checkEssayPermissions(essay, userId);
 
     await this.essayRepository.deleteEssay(essay);
@@ -495,7 +480,7 @@ export class EssayService {
 
   @Transactional()
   async saveThumbnail(file: Express.Multer.File, essayId?: number) {
-    const fileName = await this.getFileName(essayId);
+    const fileName = await this.getFileNameByThumbnail(essayId);
     const newExt = file.originalname.split('.').pop();
 
     const imageUrl = await this.awsService.imageUploadToS3(fileName, file, newExt);
@@ -506,7 +491,7 @@ export class EssayService {
   async deleteThumbnail(essayId: number) {
     const essay = await this.essayRepository.findEssayById(essayId);
     if (!essay.thumbnail) {
-      throw new NotFoundException('No thumbnail to delete');
+      throw new HttpException('삭제할 썸네일이 없습니다.', HttpStatus.NOT_FOUND);
     }
 
     const urlParts = essay.thumbnail.split('/').pop();
@@ -519,7 +504,7 @@ export class EssayService {
     return { message: 'Thumbnail deleted successfully' };
   }
 
-  private async getFileName(essayId?: number): Promise<string> {
+  async getFileNameByThumbnail(essayId?: number): Promise<string> {
     const uuid = this.utilsService.getUUID();
     if (!essayId) {
       return `images/${uuid}`;
@@ -558,10 +543,11 @@ export class EssayService {
       const recentTagObjects = await this.essayRepository.getRecentTags(recentEssayIds);
       recentTags = recentTagObjects.map((tag) => tag.tagId);
     }
+
     return recentTags;
   }
 
-  private getRandomElements<T>(array: T[], count: number): T[] {
+  getRandomElements<T>(array: T[], count: number): T[] {
     const shuffled = array.sort(() => 0.5 - Math.random());
     return shuffled.slice(0, count);
   }
@@ -599,22 +585,33 @@ export class EssayService {
   }
 
   async saveEssays(essays: Essay[]) {
-    return await this.essayRepository.saveEssays(essays);
+    return this.essayRepository.saveEssays(essays);
   }
 
   async updatedEssaysOfStory(userId: number, story: Story, reqEssayIds: number[]) {
-    const exEssayIds = story.essays.map((essay) => essay.id);
+    // 수정할 스토리에 포함된 에세이 아이디 셋
+    const essayIdsByStory = new Set(story.essays.map((essay) => essay.id));
+    // 요청으로 들어온 에세이 아이디 셋
+    const reqEssayIdsSet = new Set(reqEssayIds);
 
-    const addedEssayIds = reqEssayIds.filter((essayId) => !exEssayIds.includes(essayId));
-    const removedEssayIds = exEssayIds.filter((essayId) => !reqEssayIds.includes(essayId));
+    // 스토리에서 추가되어야 하는 에세이 아이디
+    const addedEssayIds = reqEssayIds.filter((essayId) => !essayIdsByStory.has(essayId));
+    // 스토리에서 제거되어야 하는 에세이 아이디
+    const removedEssayIds = Array.from(essayIdsByStory).filter(
+      (essayId) => !reqEssayIdsSet.has(essayId),
+    );
+
+    const operations = [];
 
     if (addedEssayIds.length > 0) {
-      await this.addEssaysStory(userId, addedEssayIds, story);
+      operations.push(this.addEssaysStory(userId, addedEssayIds, story));
     }
 
     if (removedEssayIds.length > 0) {
-      await this.deleteEssaysStory(userId, removedEssayIds);
+      operations.push(this.deleteEssaysStory(userId, removedEssayIds));
     }
+
+    await Promise.all(operations);
   }
 
   async addEssaysStory(userId: number, essayIds: number[], story: Story) {
